Change pathing of doTypesMatch for consistent index.js
